tests/wpt/tests/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-namespace.html

<!DOCTYPE html>
<title>getElementsByName and foreign namespaces</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-getelementsbyname">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test">
<p name="math"><math name="math">
<mi>a</mi>
<mo>+</mo>
<mi>b</mi>
</math>
<p name="svg"><svg width="300" height="100" name="svg">
<rect width="300" height="100" fill="rgb(0,0,255)"/>
</svg>
</div>
<script>
test(function() {
  var ps = document.getElementById("test")
                   .getElementsByTagName("p");
  assert_equals(document.getElementsByName("math").length, 1);
  assert_equals(document.getElementsByName("math")[0], ps[0]);
  assert_equals(document.getElementsByName("svg").length, 1);
  assert_equals(document.getElementsByName("svg")[0], ps[1]);
});
</script>
