LayoutTests/imported/w3c/web-platform-tests/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-namespace-expected.txt


FAIL getElementsByName and foreign namespaces assert_equals: expected 1 but got 2
a
+
b
